Add tests for current offences task
Also reorganises the existing tests to create an ‘Offence and licence information’ section.
